(feat) Transform delegate list into component
