+++ src/components/WalletButton/Alert.tsx
@@ -21,7 +21,7 @@ function getModalStyle() {
     top: `${top}%`,
     left: `${left}%`,
     transform: `translate(-${top}%, -${left}%)`,
-    ["@media (max-width:414px)"]: {
+    ["@media (maxWidth:414px)"]: {
       top: "30%",
       left: "35%",
     },
@@ -42,7 +42,7 @@ const useStyles = makeStyles((theme) => ({
     // backgroundImage: "linear-gradient(to  top, #ECF8FF, #FFF6F8)"
     boxShadow: theme.shadows[5],
     // padding: theme.spacing(2, 4, 3),
-    ["@media (max-width:414px)"]: {
+    ["@media (maxWidth:414px)"]: {
       width: 300,
     },
   },
